pyproject.toml
[project]
name = "capstonellm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.12"
dependencies = [
    "boto3>=1.36.9",
    "openai>=2.17.0",
    "pyspark>=3.5.4",
    "pytest>=8.3.4",
    "python-dotenv>=1.2.1",
    "requests>=2.32.3",
    "tenacity>=9.1.3",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
